Refactoring get_org_languages
