+++ lib/coophub/backends/github.ex
@@ -86,23 +86,6 @@ defmodule Coophub.Backends.Github do
     Repos.get_percentages_by_language(languages)
   end
 
-  @spec get_org_languages(org()) :: map
-  def get_org_languages(%Organization{:repos => repos}) do
-    languages =
-      Enum.reduce(repos, %{}, fn %Repository{:languages => langs} = repo, acc ->
-        if not repo.fork do
-          Enum.reduce(langs, acc, fn {lang, %{"bytes" => bytes}}, acc_repo ->
-            acc_lang = Map.get(acc, lang, 0)
-            Map.put(acc_repo, lang, acc_lang + bytes)
-          end)
-        else
-          acc
-        end
-      end)
-
-    Repos.get_percentages_by_language(languages)
-  end
-
   ########
   ## INTERNALS
   ########
